+++ CHANGELOG.rst
@@ -12,6 +12,22 @@ Changelog
 ==========
 
 +++++++++
+v0.3.1 (01/08/2024)
++++++++++
+
+**Fixed**
+
+- Use new proposal-specific endpoints for dewar registry data
+
++++++++++
+v0.3.0 (16/07/2024)
++++++++++
+
+**Added**
+
+- Allow new generic TLC types
+
++++++++++
 v0.2.0 (06/06/2024)
 +++++++++